refactor(frontend): replace search page, avoid useeffect and many rerenders
